requirements.txt
clickhouse-connect
pandas
boto3
psycopg2-binary
pyarrow
pyyaml
sqlalchemy
pymysql